#1473 Hopefully bring this ready for review
