+++ src/psyclone/psyir/transformations/__init__.py
@@ -40,6 +40,7 @@
 transformations and base classes.
 '''
 
+from psyclone.psyir.transformations.block_loop_trans import BlockLoopTrans
 from psyclone.psyir.transformations.extract_trans import ExtractTrans
 from psyclone.psyir.transformations.hoist_trans import HoistTrans
 from psyclone.psyir.transformations.loop_trans import LoopTrans
@@ -70,7 +71,8 @@ from psyclone.psyir.transformations.transformation_error \
 # this package e.g.:
 # from psyclone.psyir.transformations import ExtractTrans
 
-__all__ = ['ExtractTrans',
+__all__ = ['BlockLoopTrans',
+           'ExtractTrans',
            'NanTestTrans',
            'ProfileTrans',
            'PSyDataTrans',
