+++ src/psyclone/transformations.py
@@ -54,7 +54,6 @@ from fparser.two.Fortran2003 import Subroutine_Subprogram, \
 
 from psyclone import psyGen
 from psyclone.configuration import Config
-from psyclone.core import VariablesAccessInfo, Signature, AccessType
 from psyclone.domain.lfric import LFRicConstants
 from psyclone.dynamo0p3 import DynInvokeSchedule
 from psyclone.errors import InternalError
@@ -65,8 +64,7 @@ from psyclone.psyir import nodes
 from psyclone.psyir.nodes import CodeBlock, Loop, Assignment, Schedule, \
     Directive, ACCLoopDirective, OMPDoDirective, OMPParallelDoDirective, \
     ACCDataDirective, ACCEnterDataDirective, OMPDirective, \
-    ACCKernelsDirective, Routine, OMPTaskloopDirective, BlockedLoop, Literal, \
-    BinaryOperation, Reference
+    ACCKernelsDirective, Routine, OMPTaskloopDirective
 from psyclone.psyir.symbols import SymbolError, ScalarType, DeferredType, \
     INTEGER_TYPE, DataSymbol, Symbol
 from psyclone.psyir.transformations import RegionTrans, LoopTrans, \
@@ -1136,185 +1134,6 @@ class GOceanOMPLoopTrans(OMPLoopTrans):
         self.validate(node, options=options)
 
         return OMPLoopTrans.apply(self, node, options)
-
-
-class BlockLoopTrans(LoopTrans):
-    '''
-    Apply a blocking transformationt to a loop (in order to permit a
-    chunked parallelisation). For example:
-
-    TODO
-    >>>
-    '''
-    # TODO: Finish docstring
-    def __str__(self):
-        return "Split a loop into a blocked loop pair"
-
-    def validate(self, node, options=None):
-        '''
-        Validates that the Loop represented by :py:obj:`node` can have
-        a BlockLoopTrans applied.
-
-        :param node: the loop to validate.
-        :type node: :py:class:`psyclone.psyir.nodes.Loop`
-        :param options: a dictionary with options for transformation.
-        :type options: dictionary of string:values or None
-        :param int options["blocksize"]: The size to block over for this \
-                transformation. If not specified, the value 32 is used.
-
-        :raises TransformationError: if the supplied Loop has a step size \
-                which is not a constant value.
-        :raises TransformationError: if the supplied Loop has a step size \
-                larger than the chosen block size.
-        :raises TransformationError: if the supplied Loop has an ancestor \
-                blocked loop.
-        :raises TransformationError: if the supplied Loop writes to Loop \
-                variables inside the Loop body.
-        '''
-        super(BlockLoopTrans, self).validate(node, options=options)
-        # If step is a variable don't support it for now.
-        if options is None:
-            options = {}
-        if isinstance(node.children[2], nodes.Reference):
-            raise TransformationError("Cannot apply a BlockLoopTrans to "
-                                      "a loop with a non-constant step size")
-        block_size = options.get("blocksize", 32)
-        if abs(int(node.children[2].value)) > block_size:
-            raise TransformationError("Cannot apply a BlockLoopTrans to "
-                                      "a loop with larger step size than the "
-                                      "chosen block size")
-        if node.ancestor(BlockedLoop) is not None:
-            raise TransformationError("Cannot apply a BlockedLoopTrans to "
-                                      "a loop with a parent BlockedLoop node")
-        # Other checks needed for validation
-        # Dependency analysis, following rules:
-        # No child has a write dependency to the loop variable.
-        # Find variable access info for the loop variable and step
-        refs = VariablesAccessInfo(node.children[0])
-        node_vars = VariablesAccessInfo()
-        if refs is not None:
-            node_vars.merge(refs)
-        refs = VariablesAccessInfo(node.children[1])
-        if refs is not None:
-            node_vars.merge(refs)
-        # The current implementation of BlockedLoopTrans does not allow
-        # the step size to be non-constant, but we include it for future
-        # proofing.
-        refs = VariablesAccessInfo(node.children[2])
-        if refs is not None:
-            node_vars.merge(refs)
-
-        # Add the access pattern to the node variable name
-        node_vars.add_access(Signature(node.variable.name),
-                             AccessType.WRITE, self)
-        node_vars.add_access(Signature(node.variable.name),
-                             AccessType.READ, self)
-
-        #refs = VariablesAccessInfo(node.variable)
-        #if refs is not None:
-        #    node_vars.merge(refs)
-        node_sigs = node_vars.all_signatures
-
-        #Find the Loop code's signatures
-        child_refs = VariablesAccessInfo(node.children[3])
-        child_sigs = child_refs.all_signatures
-
-        for ref1 in node_sigs:
-            if ref1 not in child_sigs:
-                continue
-            access1 = node_vars[ref1]
-            access2 = child_refs[ref1]
-
-            # If access2 is a write then we write to a loop variable
-            if access2.is_written():
-                raise TransformationError("Cannot apply a BlockedLoopTrans "
-                                          "to a loop where loop variables are "
-                                          "written to inside the loop body.")
-
-
-    def apply(self, node, options=None):
-        '''
-        Converts the Loop represented by :py:obj:`node` into a
-        nested loop where the outer loop is over blocks and the inner
-        loop is over each block.
-
-        :param node: the loop to transform.
-        :type node: :py:class:`psyclone.psyir.nodes.Loop`
-        :param options: a dictionary with options for transformations.
-        :type options: dictionary of string:values or None
-        :param int options["blocksize"]: The size to block over for this \
-                transformation. If not specified, the value 32 is used.
-
-        :returns: Tuple of None and None
-        :rtype: (None, None)
-        '''
-
-        self.validate(node, options)
-        if options is None:
-            options = {}
-        block_size = options.get("blocksize", 32)
-        schedule = node.root
-
-        # Create (or find) the symbols we need for the blocking transformation
-        routine = node.ancestor(nodes.Routine)
-        end_inner_loop = routine.symbol_table.symbol_from_tag(
-                "el_inner", symbol_type=DataSymbol,
-                datatype=node.variable.datatype)
-        outer_loop_variable = routine.symbol_table.symbol_from_tag(
-                "out_var", symbol_type=DataSymbol,
-                datatype=node.variable.datatype)
-        end_outer_loop = routine.symbol_table.symbol_from_tag(
-                "el_outer", symbol_type=DataSymbol,
-                datatype=node.variable.datatype)
-
-        # TODO: Check if any ancestors use these variables?
-
-        # Store the node's parent for replacing later and the start and end
-        # indicies
-        loop_parent = node.parent
-        c0 = node.children[0]
-        c1 = node.children[1]
-
-        # Create the end bound for the inner loop
-        inner_loop_end = None
-        inner_loop_end_lit = Reference(end_inner_loop)
-        # For positive steps we do el_inner = min(out_var+block_size, el_outer)
-        # For negative steps we do el_inner = max(out_var-block_size, el_outer)
-        if int(node.children[2].value) > 0:
-            add = BinaryOperation.create(BinaryOperation.Operator.ADD,
-                                         Reference(outer_loop_variable),
-                Literal("{0}".format(block_size), node.variable.datatype))
-            mn = BinaryOperation.create(BinaryOperation.Operator.MIN, add, 
-                                        c1.copy())
-            inner_loop_end = Assignment.create(Reference(end_inner_loop), mn)
-        elif int(node.children[2].value) < 0:
-            sub = BinaryOperation.create(BinaryOperation.Operator.SUB, 
-                                         Reference(outer_loop_variable),
-                Literal("{0}".format(block_size), node.variable.datatype))
-            mx = BinaryOperation.create(BinaryOperation.Operator.MAX, sub,
-                                        c1.copy())
-            inner_loop_end = Assignment.create(Reference(end_inner_loop), mx)
-            # block_size needs to be negative if we're reducing
-            block_size = -block_size
-        else:
-            # step is 0 so assert for now
-            assert False
-
-        # Replace the inner loop start and end with the blocking ones
-        c0.replace_with(Reference(outer_loop_variable))
-        c1.replace_with(Reference(end_inner_loop))
-
-        # Create the outerloop
-        outerloop = BlockedLoop.create(outer_loop_variable, c0, c1,
-                                       Literal("{0}".format(block_size),
-                                               outer_loop_variable.datatype),
-                                       [inner_loop_end])
-        # Replace this loop with the outerloop
-        node.replace_with(outerloop)
-        # Add the loop to the innerloop's schedule
-        outerloop.children[3].addchild(node)
-
-        return None, None
 
 
 class ColourTrans(LoopTrans):
